requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-countries==7.5.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.21
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
